fix: rename CLI marshal → conduct (package, module, command, prog name)
